// packages/client/src/types/Submit.ts
/**
 * Copyright (c) Formium, Inc. and its affiliates.
 *
 * This source code is licensed under the Business Source License found in the
 * LICENSE file in the root directory of this source tree.
 */
/**
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 *
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/* eslint-disable @typescript-eslint/consistent-type-assertions */
export interface Submit {
  /**
   * Unique identifier of submission.
   */
  id: string;

  /**
   * The form that the submission belongs to.
   */
  formId: string;

  /**
   * The project that the submission belongs to.
   */
  projectId: string;

  /**
   * The data that was submitted.
   */
  data: { [key: string]: any };

  /**
   * IP address of the submission.
   */
  ip: string;

  /**
   * Created timestamp.
   */
  createAt: string;

  /**
   * Deleted timestamp.
   */
  deleteAt?: string;

  /**
   * Deleted by user identifier.
   */
  deleteId?: string;
}